Foundry_Rag/rag_agent/requirements.txt
azure-ai-projects>=2.0.0
azure-identity
python-dotenv
openai 
tiktoken
azure.ai.projects
azure.ai.projects.models